Error control for documents and programs

+++ app/src/main/java/com/wfcrc/utils/VolleyCallback.java
@@ -3,11 +3,14 @@ package com.wfcrc.utils;
 import android.content.Context;
 import android.util.Log;
 
+import com.wfcrc.pojos.Document;
 import com.wfcrc.repository.JSONDocumentRepository;
+import com.wfcrc.repository.RepositoryException;
 import com.wfcrc.sqlite.WFCRCDB;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.List;
 
 /**
  * Created by maria on 10/26/16.
@@ -34,10 +37,17 @@ public class VolleyCallback {
 
     public  void onResponse(String response){
         Log.d("Volley document request","Response is: "+ response);
-        //store documents in database
-        WFCRCDB db = new WFCRCDB(mContext);
-        db.deleteDocuments();
-        db.insertDocuments(JSONDocumentRepository.getAll(response));
+        try {
+            List<Document> documents = JSONDocumentRepository.getAll(response);
+            if (!documents.isEmpty()) {
+                //store documents in database
+                WFCRCDB db = new WFCRCDB(mContext);
+                db.deleteDocuments();
+                db.insertDocuments(documents);
+            }
+        } catch (RepositoryException e) {
+            e.printStackTrace();
+        }
     }
 
     public void onError(String error){
